==> R/03_ICBBVA_ready_df_shp_to_plot.Rmd <==
all_zipcodes for plotting
========================================================

Get the all_zipcodes data. 
Get ready the shapefiles to draw them with ggplot2.


0. wd and libraries
```{r}
setwd("~/InnovaChallengeBBVA/R")

library(rgeos)
library(maptools)
library(sp)
library(rgdal)
library(ggplot2)

options(stringsAsFactors = FALSE)
```


1. Read the data

```{r read_the_data}
all_zipcodes <- read.csv("rawdata/all_zipcodes.csv", header = TRUE, colClasses = "character")

# Switch the variable to the right class
all_zipcodes$num_cards <- as.integer(all_zipcodes$num_cards)
all_zipcodes$num_payments <- as.integer(all_zipcodes$num_payments)
all_zipcodes$incomes <- as.numeric(all_zipcodes$incomes)
								   
# Load the metadata
dates <- read.csv("metadata/dates.csv", stringsAsFactors = FALSE)
postal_codes <- read.csv("metadata/postal_codes.csv", stringsAsFactors = FALSE)
categories <- read.csv("metadata/merchants_categories.csv", stringsAsFactors = FALSE) 
```

2. Subset only the rows where the origon cp is in Madrid Capital.

```{r subset_cp_origin_Madrid}
all_zipcodes_mad <- all_zipcodes[all_zipcodes$cp_origin %in% postal_codes$label, ]
```

3. Sum per month (to see the whole period, selection by month won't be possible)
```{r aggregate_month_and_cp_origin}
# Aggregate by month

no_month <- all_zipcodes_mad

no_month <- aggregate(cbind(num_cards, num_payments, incomes) ~ cp_origin + cp_dest + category, data = no_month, sum)
```


4. Add a incomes percentage column. 
> For every category
> For every cp_origin
> Calculate the percentages

```{r incomes_percentages}
expenditure <- no_month
expenditure_per <- data.frame()

for (cat in categories$code){
	exp_cat <- expenditure[expenditure$category == cat, ]
	for (cp in postal_codes$label){
		exp_cat_cp <- exp_cat[exp_cat$cp_origin == cp, ]
		exp_cat_cp$incomes_per <- round(prop.table(exp_cat_cp$incomes)*100, digits = 2)
		expenditure_per <- rbind(expenditure_per, exp_cat_cp)
	}
}

rm(exp_cat, exp_cat_cp, cat, cp)

# test
length(expenditure_per$cp_origin)
length(expenditure$cp_origin)
```

5. Add 'comercios' column

> Take the matrix cp and num comercios from the poly data
> add to every cp_dest the correspondant number of 'comercios'


```{r comercios_column}

# Get in which 'cp_poly_mad' polygons fall the 'comerciales' points and name them as 'comercial'

poly_comerciales <- over(as(comerciales, "SpatialPoints"), as(cp_poly_mad, "SpatialPolygons")) # needs to be 'SpatialPoints' and 'SpatialPolygons' class

poly_comerciales
# NA indicates that the point does not fall in any polygon.
# The number indicates the is the index (number) of the polygon in which a point falls (these will be the 'comerciales')

poly_comerciales <- poly_comerciales-1 # The IDs in cp_poly_mad start in 0.

## number of shops in each cp
freq <- as.data.frame(table(poly_comerciales), stringsAsFactors = FALSE)
freq
# length = 39 indicates that there are 39 polygons with at least one comerce

# merge the freq with the cp_poly_mad@data 
cp_poly_mad$id <- row.names(cp_poly_mad@data)
colnames(freq)[names(freq) == "poly_comerciales"] <- "id"


cp_poly_mad@data <- merge(cp_poly_mad@data, freq, all = TRUE)
cp_poly_mad$Freq[is.na(cp_poly_mad$Freq)] <- 0
cp_poly_mad$commercial <- cut(cp_poly_mad$Freq, 
								 breaks = c(0, 1, 4, 7, 10,Inf), 
								 labels = c("0 stores", "1 - 3 stores", "4 - 6 stores", "7 - 10 stores", "11 stores"), 
								 right = FALSE)

cp_poly_mad@data <- cp_poly_mad@data[order(cp_poly_mad$GEOCODIGO, decreasing = FALSE), ]


# subset the cp_poly_mad@data cp to get the 'num_comercios' matrix
n_comercios <- subset(cp_poly_mad@data, select = c(GEOCODIGO, Freq))

# Assign them to the correspondant cp_dest in the expenditure_per df
colnames(n_comercios)[names(n_comercios) == "GEOCODIGO"] <- "cp_dest"

expenditure_per <- merge(expenditure_per, n_comercios)

rm(freq, n_comercios)

```

6. Add origin(lon, lat) and dest (long lat) columns (4 new columns required)

```{r add_origin_dest_lon_lat}
# Get the centroids of every polygon

coord <- as.data.frame(coordinates(cp_poly_mad))

colnames(coord) <- c("lon", "lat")
coord$id <- row.names(coord)

centroids <- merge(cp_poly_mad@data[ ,c(1,3)], coord, by = "id")

colnames(centroids) <- c("id", "cp_origin","lon_origin", "lat_origin")
expenditure_per_origin <- merge(expenditure_per, centroids, by = "cp_origin")

colnames(centroids) <- c("id", "cp_dest","lon_dest", "lat_dest")
expenditure_per <- merge(expenditure_per_origin, centroids[ ,c(2:4)], by = "cp_dest")

write.csv(expenditure_per, "metadata/expenditure_per.csv", row.names = FALSE)
```

9. Shapefiles: Fortify the cp_poly_mad (with the Freq added)
```{r fortify_polygons} 

data_cpPolyMad <- fortify(cp_poly_mad, region = "id")

data_cpPolyMad <- merge(data_cpPolyMad, cp_poly_mad@data, by = "id") 
str(data_cpPolyMad)
data_cpPolyMad$commercial <- as.character(data_cpPolyMad$commercial)
```
