==> R/05_ICBBVA_loop_plot_bar.Rmd <==
04. Loop to plot the bar charts
========================================================


```{r needs}
library(reshape2)
library(ggplot2)
library(grid)

expenditure_per <- read.csv("metadata/expenditure_per.csv")
```


1. cut the freq variable

```{r}
# Cut the Freq variable to have 4 levels of comercios (0,1,2,3,4). La freq viene del 03, linea 97

expenditure_per$commercial <- cut(expenditure_per$Freq, 
								 breaks = c(0, 1, 4, 7, 10,Inf), 
								 labels = c("0 stores", "1 - 3 stores", "4 - 6 stores", "7 - 10 stores", "11 stores"), 
								 right = FALSE)

```

2. Loops

```{r}

valid_columns <- c("cp_dest", "cp_origin", "category", "incomes_per", "contiguous", "commercial")

expenditure_bars <- expenditure_per[ ,names(expenditure_per) %in% valid_columns]

for (cat in categories$code){
	cat_label <- categories$description[categories$code == cat]
	if ((cat %in% expenditure_bars$category) == FALSE){
		print(cat)
		} else {
			temp_cat <- expenditure_bars[expenditure_bars$category == cat, ]
			
			for (cp in postal_codes$label){
				if ((cp %in% temp_cat$cp_origin) == FALSE){
					print(cat)
					print(cp)
				} else {
					temp_cat_cp <- temp_cat[temp_cat$cp_origin == cp, ]
					temp_cat_cp$commercial <- as.character(temp_cat_cp$commercial)
					
					#temp_cat_cp$commercial[temp_cat_cp$contiguous == "exacto"] <- "exact" 
					temp_cat_cp$commercial[temp_cat_cp$cp_dest == cp] <- "exact"
					
					bars_data <- aggregate(incomes_per ~ category + cp_origin + commercial, data = temp_cat_cp, sum)
					
					commercial <- c("exact", "0 stores", "1 - 3 stores", "4 - 6 stores", "7 - 10 stores", "11 stores")
					
					for (com in commercial){
						if ((com %in% bars_data$commercial) == FALSE){
						missing <- c(cat, cp, com, 0)
						bars_data <- rbind(bars_data, missing)
						}
					}	
				
# 					bars_data$commercial <- factor(bars_data$commercial, levels = c("exact", "0 stores", "1 - 3 stores", "4 - 6 stores", "7 - 10 stores", "11 stores"))
					bars_data$commercial <- factor(bars_data$commercial, levels = c("11 stores", "7 - 10 stores", "4 - 6 stores", "1 - 3 stores", "0 stores", "exact"))
					
					bars_data$incomes_per <- as.numeric(bars_data$incomes_per)
					
					# Plot the bar
					ggplot() +
						geom_bar(data = bars_data, aes(x = commercial, y = incomes_per, fill = commercial), stat = "identity") +
						coord_flip() +
						scale_fill_manual(values = c("exact" = "#ebb83b", "0 stores" = "#99E4DB", "1 - 3 stores" = "#7DC7BE", "4 - 6 stores" = "#6FB9B0", "7 - 10 stores" = "#61AAA1", "11 stores" = "#539C93")) +
						geom_text(data = bars_data, aes(x= commercial, y=incomes_per+5, label=paste(round(incomes_per, digits= 0), "%", sep = "")), size=9, colour = "#332f2e", family="Courier", fontface="bold") +
						scale_y_continuous(limits=c(0, 100)) +
						ggtitle(paste("Expenditure in / out the", cp, "postal code\nfor the category", cat_label, sep = " ")) +
					theme(legend.position = "none", 
						  panel.margin = unit(20, "lines"),
						  panel.background = element_blank(),
						  panel.grid.minor = element_blank(),
						  axis.ticks = element_blank(),
						  axis.title = element_blank(),
						  axis.text = element_blank(),
						  plot.title = element_text(lineheight=0.8, family="Courier", size = 20,  vjust=2), 
						  panel.grid.major.x = element_line(size = .6, colour = "#c4cac9", linetype = "dashed"), 
						  panel.grid.major.y = element_blank())
					
					# Save the plot
					print(last_plot())
					filename <- paste(getwd(), "/graphs/bar_", cp, "_", cat, ".png", sep = "")
					dev.copy(file = filename, device=png, bg="white", width=1000, height=1000) 
				#dev.copy(file = "bar.png", device=png, bg="white",  width=1000, height=1000) 
    				dev.off()
				}
			}
		}
}
```
